Splits classes into new files in tests
