+++ tests/SpacePadTransformer.php
@@ -1,0 +1,11 @@
+<?php
+
+use App\ContentParser\Transformer;
+
+class SpacePadTransformer implements Transformer
+{
+    public function transform($content)
+    {
+        return " {$content} ";
+    }
+}
