+++ tests/TrimTransformer.php
@@ -1,0 +1,11 @@
+<?php
+
+use App\ContentParser\Transformer;
+
+class TrimTransformer implements Transformer
+{
+    public function transform($content)
+    {
+        return trim($content);
+    }
+}
